requirements.txt
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0
scipy==1.9.3
seaborn==0.12.1
torch==1.13.0
torchvision==0.14.0
